pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vulnscan"
version = "0.1.0"
description = "Web application vulnerability scanner"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-timeout>=2.2"]

[project.scripts]
vulnscan = "vulnscan.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
timeout = 60
markers = [
    "integration: end-to-end tests (full scan, slow)",
]
